requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
openai
google-generativeai
pyswisseph
opik
